+++ NEWS.md
@@ -6,11 +6,15 @@ now denotes the lowest possible grouping unit, i.e. the unit of one observation
 and needs to be specified explicitly. The default is now `NULL`. The reason for
 this change is that most metrics need scoring on the observation level and this 
 the most consistent implementation of this principle. The pit function receives
-its grouping now from `summarise_by`. 
-- (potentially planned) rename true_values to true_value and predictions to prediction
+its grouping now from `summarise_by`. In a similar spirit, `summarise_by` has to
+be specificed explicitly and e.g. doesn't assume anymore that you want 'range'
+to be included. 
+- for the interval score, `weigh = TRUE` is now the default option.
+- (potentially planned) rename true_values to true_value and predictions to prediction. 
 
 ### Feature updates
-- updated quantile evaluation metrics in `eval_forecasts`
+- updated quantile evaluation metrics in `eval_forecasts`. Bias as well as 
+calibration now take all quantiles into account
 - Included option to summarise scores according to a `summarise_by` argument in 
 `eval_forecasts` The summary can return the mean, the standard deviation as well
 as an arbitrary set of quantiles.